src/stolon/openapi_generated/billing_event_dev/pyproject.toml
[tool.poetry]
name = "open-api-definition-client"
version = "v0"
description = "A client library for accessing OpenAPI definition"
authors = []
readme = "README.md"
packages = [
    { include = "open_api_definition_client" },
]
include = ["CHANGELOG.md", "open_api_definition_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
